requirements.txt
anthropic
python-slugify
markdown
